dude_ml/requirements.txt
pandas
requests
biopython
jupyter
scikit-learn
matplotlib
seaborn
rdkit
lightgbm
tqdm
useful_rdkit_utils
